+++ generators/premiums/gpt3_grammar_correction/__init__.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 INPUT_EXAMPLE = {
     "apiKey": "<API_KEY_GOES_HERE",
     "prompt": "Named must be your fear before banish it you can.",
-    "temperature": 0,
+    "temperature": 0.0,
     "maxTokens": 64,
     "top_p": 1.0,
     "frequencyPenalty": 0.0, 
@@ -14,7 +14,7 @@ INPUT_EXAMPLE = {
 class Gpt3GrammarCorrectionModel(BaseModel):
     apiKey: str
     prompt: str
-    temperature: int
+    temperature: float
     maxTokens: int
     top_p: float
     frequencyPenalty: float
